+++ packages/react-native/src/internal/fcmContract.d.ts
@@ -35,6 +35,8 @@ export type NotifyKitAndroidConfig = {
   pressAction?: NotifyKitPressAction;
   actions?: NotifyKitAndroidAction[];
   style?: NotifyKitAndroidStyle;
+  showTimestamp?: boolean;
+  timestamp?: number;
 };
 
 export type NotifyKitIosAttachment = {